+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+xxx
+===================
+- Fixed `RequestQueue.addRequest()` to gracefully handle invalid URLs
+- Renamed `RequestOperationInfo` to `QueueOperationInfo`
+- Added `request` field to `QueueOperationInfo`
+
 0.8.18 / 2018-10-30
 ===================
 - Fix a bug in `CheerioCrawler` where `useApifyProxy` would only work with `apifyProxyGroups`.